Conflicts:
	include/media/mediarecorder.h
	include/media/stagefright/MetaData.h
	include/media/stagefright/OMXCodec.h
	media/libmediaplayerservice/Android.mk
	media/libmediaplayerservice/StagefrightRecorder.cpp
	media/libstagefright/Android.mk
	media/libstagefright/DataSource.cpp
	media/libstagefright/MPEG4Extractor.cpp
	media/libstagefright/OMXCodec.cpp
	media/libstagefright/SampleIterator.cpp
	media/libstagefright/SampleTable.cpp
	media/libstagefright/StagefrightMediaScanner.cpp
	media/libstagefright/httplive/LiveSession.cpp
	media/libstagefright/include/LiveSession.h

+++ media/libstagefright/SampleIterator.cpp
@@ -305,7 +305,8 @@ status_t SampleIterator::findSampleTime(
         ++mTimeToSampleIndex;
     }
 
-    *time = mTTSSampleTime + (uint64_t)mTTSDuration * (sampleIndex - mTTSSampleIndex);
+    *time = mTTSSampleTime + (uint64_t)mTTSDuration * ((uint64_t)sampleIndex - (uint64_t)mTTSSampleIndex);
+
     *time += (int32_t)mTable->getCompositionTimeOffset(sampleIndex);
 
     return OK;
